buffered output to improve performance

+++ src/buffer.rs
@@ -31,14 +31,14 @@ impl Buffer {
         self.lines.len()
     }
 
-    pub fn insert(&mut self, x: u16, y: usize, c: char) {
+    pub fn insert(&mut self, x: usize, y: usize, c: char) {
         if let Some(line) = self.lines.get_mut(y) {
             (*line).insert(x as usize, c);
         }
     }
 
     /// removes a character from the buffer
-    pub fn remove(&mut self, x: u16, y: usize) {
+    pub fn remove(&mut self, x: usize, y: usize) {
         if let Some(line) = self.lines.get_mut(y) {
             (*line).remove(x as usize);
         }
